+++ app/changelog/changelog.ts
@@ -15,6 +15,7 @@ export const changelog: ChangelogItem[] = [
     fixed: [
       'Besonders lange Wörter werden nun auf mehrere Zeilen umgebrochen, statt über den Bildrand hinauszuragen.',
       'Beim Klick auf eine private Nachricht in der Sidebar wird nun die Sidebar geschlossen.',
+      'HTML und Emojis in Privatnachrichten werden nun korrekt geparsed.',
     ],
   },
   {